requirements.txt
python_gitlab==4.1.1
pysnyk==0.9.18